requirements.txt
spotipy==2.22.1
slskd-api==0.1.0
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
fuzzywuzzy==0.18.0
rapidfuzz==3.5.2